Changed nClasses to 3.
After trying the code again this morning, I noticed that there was a problem with the input of the data from Cthulhu.txt to the final DNN. I reasoned out that it was due to the number of classes, and an incorrect call to a non-training document. I manually set the classes to 3 in nClasses to rectify this problem.